fix #834: avoid warnings in Python 3.8
* use == to compare to a string literal instead of "is"
* reworked sequence() to avoid comparing mult to 1 using "is"
* moved compatibility layer with python2.7 to a specific compat module and import abstract base classes from there
