+++ larray/inout/misc.py
@@ -1,10 +1,10 @@
 from __future__ import absolute_import, print_function
 
-import numpy as np
 from pandas import DataFrame
 
 from larray.core.constants import nan
-from larray.util.misc import StringIO, deprecate_kwarg
+from larray.util.misc import deprecate_kwarg
+from larray.util.compat import StringIO
 from larray.inout.common import _get_index_col
 from larray.inout.pandas import df_asarray
 from larray.inout.csv import read_csv
